rename tab files to avoid ambiguity with diagnostics page
